m3u8_downloader/requirements.txt
requests
pycrypto==2.6.1
